pyproject.toml
[project]
name = "ai-lecture-4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.2.1",
    "jsonschema>=4.24.0",
]
